requirements.txt
vt-py
Pillow
